Fix save/load bugs

+++ src/LoadSaveData.py
@@ -11,117 +11,61 @@ from PyQt5.QtWidgets import QMessageBox, QFileDialog
 
 # Loads settings
 ################################################################################
-def loadSettings():
+def loadSettings(userLoad=False):
     settings = {}
-    settingsDir = 'Settings.json'
+    settingsPath = '../data/Settings.json'
 
-    if not os.path.exists(settingsDir):
-        msg = QMessageBox()
-        msg.setText('No settings file found.')
-        msg.setInformativeText('What would you like to do?')
-        msg.setStandardButtons(QMessageBox.Open | QMessageBox.Ignore)
-        returnVal = msg.exec_()
-
-        # Browse for file
-        if returnVal == QMessageBox.Open:
-            fileName = QFileDialog.getOpenFileName(self,
-                'Browse for Settings.json', '.', 'Settings.json')
-            if os.path.exists(fileName[0]):
-                with open(fileName[0], 'r') as f:
-                    settings = json.load(f)
-                settingsDir = fileName[0]
-            else:
-                self.settings = {
-                    'encrypted':  False,
-                    'password':   ''
-                }
-        else:
-            settings = {
-                'encrypted':  False,
-                'password':   ''
-            }
+    if not os.path.exists(settingsPath):
+        settings = {
+            'encrypted':  False,
+            'password':   ''
+        }
     else:
-        with open(settingsDir, 'r') as f:
+        with open(settingsPath, 'r') as f:
             settings = json.load(f)
 
-    return settings, settingsDir
+    return settings, settingsPath
 
 # Saves settings to file
 ################################################################################
-def saveSettings(settings, settingsDir):
-    if not settingsDir:
-        settingsDir = 'Settings.json'
+def saveSettings(settings, settingsPath):
+    if not os.path.exists(settingsPath):
+        settingsPath = '../data/Settings.json'
 
-    with open(settingsDir, 'w') as f:
+    with open(settingsPath, 'w') as f:
         json.dump(settings, f)
 
 # Loads last used account data from Accounts.json file
 ################################################################################
-def loadAccounts(settings, password):
+def loadAccounts(settings, password, userLoad=False):
     accounts = []
-    accountsDir = ''
+    accountsPath = ''
 
-    # No account files found. Browse or create new files
-    if (not settings['encrypted'] and not os.path.exists('Accounts.json')):
-        msg = QMessageBox()
-        msg.setText('No Accounts.json file found.')
-        msg.setInformativeText('What would you like to do?')
-        msg.setStandardButtons(QMessageBox.Open | QMessageBox.Ignore)
-        returnVal = msg.exec_()
-
-        # Browse for file
-        if returnVal == QMessageBox.Open:
-            fileName = QFileDialog.getOpenFileName(self,
-                'Browse for Accounts.json', '.', 'Accounts.json')
-            if os.path.exists(fileName[0]):
-                with open(fileName[0], 'r') as f:
-                    accounts = json.load(f)
-                accountsDir = fileName[0]
-            else:
-                return
-        else:
-            return
-    # No account files found. Browse or create new files
-    elif settings['encrypted'] and not os.path.exists('Accounts.enc'):
-        msg = QMessageBox()
-        msg.setText('No Accounts.enc file found.')
-        msg.setInformativeText('What would you like to do?')
-        msg.setStandardButtons(QMessageBox.Open | QMessageBox.Ignore)
-        returnVal = msg.exec_()
-
-        # Browse for file
-        if returnVal == QMessageBox.Open:
-            fileName = QFileDialog.getOpenFileName(self,
-                'Browse for Accounts.enc', '.', 'Accounts.enc')
-            if os.path.exists(fileName[0]):
-                decryptFile(password, fileName[0])
-                newFileName = os.path.dirname(os.path.abspath(fileName[0]))
-                newFileName = newFileName + '/Accounts.json'
-                with open(newFileName, 'r') as f:
-                    accounts = json.load(f)
-                accountsDir = newFileName
-            else:
-                return
-        else:
-            return
-    # Load accounts
-    elif not settings['encrypted'] and os.path.exists('Accounts.json'):
-        with open('Accounts.json', 'r') as f:
-            accounts = json.load(f)
-    # Decrypt
-    elif settings['encrypted'] and os.path.exists('Accounts.enc'):
-        decryptFile(password, 'Accounts.enc')
-        with open('Accounts.json', 'r') as f:
-            accounts = json.load(f)
-    # Else ???
+    # If encrypted
+    if not settings['encrypted']:
+        # If not a user-specified file path
+        if not userLoad:
+            accountsPath = '../data/Accounts.json'
+        # else:
+            # Set user-specified file path
+        # Open and load
+        if os.path.exists(accountsPath):
+            with open(accountsPath, 'r') as f:
+                accounts = json.load(f)
     else:
-        print('ERROR: Account files missing.')
-        return
+        # If not a user-specified file path
+        if not userLoad:
+            accountsPath = '../data/Accounts.enc'
+        # else:
+            # Set user-specified file path
+        # Decrypt, open, load, encrypt
+        if os.path.exists(accountsPath):
+            decryptFile(password, accountsPath)
+            with open('../data/Accounts.json', 'r') as f:
+                accounts = json.load(f)
+            encryptFile(password, accountsPath)
 
-    # Save and re-encrypt if necessary
-    saveAccounts(accounts, accountsDir, settings, password)
-
-    return accounts, accountsDir
+    return accounts, accountsPath
 
 # Returns last used account
 ################################################################################
@@ -136,12 +80,19 @@ def getLastUsedAccount(accounts):
 
 # Saves accounts to file
 ################################################################################
-def saveAccounts(accounts, accountsDir, settings, password):
-    if not accountsDir:
-        accountsDir = 'Accounts.json'
-    with open(accountsDir, 'w') as f:
-        json.dump(accounts, f)
-
-    # Encrypt data
+def saveAccounts(accounts, accountsPath, settings, password):
     if settings['encrypted']:
-        encryptFile(password, accountsDir)
+        if not os.path.exists(accountsPath):
+            print('ERROR:  Account file is missing. Cannot save.')
+            return
+        else:
+            decryptFile(password, accountsPath)
+            with open(accountsPath, 'w') as f:
+                json.dump(accounts, f)
+            encryptFile(password, accountsPath)
+    else:
+        if not os.path.exists(accountsPath):
+            accountsPath = '../data/Accounts.json'
+        with open(accountsPath, 'w') as f:
+            json.dump(accounts, f)
+
